+++ packages/expo-router/src/react-navigation/core/createNavigationContainerRef.tsx
@@ -19,6 +19,7 @@ export function createNavigationContainerRef<
     'dispatch',
     'dispatchSync',
     'isFocused',
+    'canDismiss',
     'canGoBack',
     'getRootState',
     'getState',